+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## New in 1.4.3
+
+### Changes
+
+- Added Base Save in Tweaks and config option to remove character level from saving throw calculations. This should make it easier to use a custom saving throw formula, albeit without much automation.
+- Added config option to switch attribute modifier breakpoints to B/X standards instead of WWN.
+- Added support for a house rule to replace System Strain with the Injuries/Wounds system from the Goblin Punch blog. If active, the strain entry on the sheet is replaced with a tracker for Injuries and Wounds. Injuries rolls are also automated if the option is active, though there is no automation of effects.
+
 ## New in 1.4.2
 
 ### Fixes